requirements.txt
Flask
requests
pycryptodome
gunicorn
